pyproject.toml
[project]
name = "ta-quarterly-report"
version = "0.1.0"
description = "Automated Technical Architect quarterly impact report generator"
requires-python = ">=3.11"
dependencies = [
    "python-pptx>=1.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
ta-report = "src.main:main"

[tool.uv]
python-preference = "managed"
